[Refactor] CBR (#404)
[Refactor] CBR

Move all API methods and related structures to a separate file.
Inline all Extracts and use the extract package instead.
Remove all Builders
