+++ openstack/cbr/v3/backups/delete.go
@@ -1,0 +1,8 @@
+package backups
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("backups", id), nil)
+	return
+}
