+++ openstack/cbr/v3/vaults/delete.go
@@ -1,0 +1,8 @@
+package vaults
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("vaults", id), nil)
+	return
+}
