+++ openstack/cbr/v3/vaults/results.go
@@ -1,23 +1,8 @@
 package vaults
 
 import (
-	"fmt"
-
-	"github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/common/tags"
 )
-
-type vaultResult struct {
-	golangsdk.Result
-}
-
-type CreateResult struct {
-	vaultResult
-}
-
-type GetResult struct {
-	vaultResult
-}
 
 type Billing struct {
 	Allocated       int    `json:"allocated"`
@@ -64,81 +49,4 @@ type Vault struct {
 	UserID     string         `json:"user_id"`
 	CreatedAt  string         `json:"created_at"`
 	AutoExpand bool           `json:"auto_expand"`
-}
-
-func (r vaultResult) Extract() (*Vault, error) {
-	var s struct {
-		Vault *Vault `json:"vault"`
-	}
-	err := r.ExtractInto(&s)
-	return s.Vault, err
-}
-
-type DeleteResult struct {
-	golangsdk.ErrResult
-}
-
-type AssociateResourcesResult struct {
-	golangsdk.Result
-}
-
-func (r AssociateResourcesResult) Extract() ([]string, error) {
-	var s struct {
-		AddResourceIDs []string `json:"add_resource_ids"`
-	}
-	if r.Err != nil {
-		return nil, r.Err
-	}
-	err := r.ExtractInto(&s)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract Associated Resource IDs")
-	}
-	return s.AddResourceIDs, nil
-}
-
-type DissociateResourcesResult struct {
-	golangsdk.Result
-}
-
-func (r DissociateResourcesResult) Extract() ([]string, error) {
-	var s struct {
-		RemoveResourceIDs []string `json:"remove_resource_ids"`
-	}
-	if r.Err != nil {
-		return nil, r.Err
-	}
-	err := r.ExtractInto(&s)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract Dissociated Resource IDs")
-	}
-	return s.RemoveResourceIDs, nil
-}
-
-type BindPolicyResult struct {
-	golangsdk.Result
-}
-
-type PolicyBinding struct {
-	VaultID  string `json:"vault_id"`
-	PolicyID string `json:"policy_id"`
-}
-
-func (r BindPolicyResult) Extract() (*PolicyBinding, error) {
-	var s struct {
-		PolicyBinding *PolicyBinding `json:"associate_policy"`
-	}
-	err := r.ExtractInto(&s)
-	return s.PolicyBinding, err
-}
-
-type UnbindPolicyResult struct {
-	golangsdk.Result
-}
-
-func (r UnbindPolicyResult) Extract() (*PolicyBinding, error) {
-	var s struct {
-		PolicyBinding *PolicyBinding `json:"dissociate_policy"`
-	}
-	err := r.ExtractInto(&s)
-	return s.PolicyBinding, err
 }
