+++ openstack/cbr/v3/vaults/unbind_policy.go
@@ -1,0 +1,23 @@
+package vaults
+
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
+)
+
+func UnbindPolicy(client *golangsdk.ServiceClient, vaultID string, opts BindPolicyOpts) (*PolicyBinding, error) {
+	b, err := golangsdk.BuildRequestBody(opts, "")
+	if err != nil {
+		return nil, err
+	}
+
+	raw, err := client.Post(client.ServiceURL("vaults", vaultID, "dissociatepolicy"), b, nil, &golangsdk.RequestOpts{
+		OkCodes: []int{200},
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	var res PolicyBinding
+	return &res, extract.IntoStructPtr(raw.Body, &res, "dissociate_policy")
+}
